Touch Bug 1243129: Mixed bag of JS cleanups/updates #854
- Fix for js container widget test
- Separate js widget tests
- Separate js utils test suites
- Add js NotImplementedError
- Add js datetime formatter util
- Add js datetime util tests
