+++ pontoon/static/js/errors/index.js
@@ -1,0 +1,10 @@
+
+export class NotImplementedError extends Error {
+
+    constructor(...args) {
+        super(...args)
+        Error.captureStackTrace(this, NotImplementedError)
+    }
+}
+
+NotImplementedError.prototype = Error.prototype;
