+++ tests/js/utils/ajax.js
@@ -1,6 +1,5 @@
 
 import {DjangoAjax, PontoonDjangoAjax, ajax} from 'utils/ajax';
-import {Strip, strip} from 'utils/strip';
 
 
 test('PontoonDjangoAjax instance', () => {
@@ -194,51 +193,4 @@ test('PontoonDjangoAjax appendParams', () => {
     expect(_ajax.appendParams(container, {foo: 3, bar: [1, 3, 7]})).toBe(container)
     expect(container.append.mock.calls).toEqual(
         [["foo", 3], ["bar", 1], ["bar", 3], ["bar", 7]])
-})
-
-
-test('Strip instance', () => {
-    expect(strip instanceof Strip).toBe(true);
-})
-
-
-test('Strip lstrip', () => {
-    const _strip = new Strip();
-    expect(_strip.lstrip('aaaxyz', 'a')).toBe('xyz');
-    expect(_strip.lstrip('aaaxyz', 'b')).toBe('aaaxyz');
-    expect(_strip.lstrip('abaxyz', 'b')).toBe('abaxyz');
-    expect(_strip.lstrip('bbaxyz', 'b')).toBe('axyz');
-    expect(_strip.lstrip('bbaxyzbbb', 'b')).toBe('axyzbbb');
-    expect(_strip.lstrip('   bbaxyz')).toBe('bbaxyz');
-    expect(_strip.lstrip('   bb   axyz')).toBe('bb   axyz');
-    expect(_strip.lstrip('   bb   axyz   ')).toBe('bb   axyz   ');
-})
-
-
-test('Strip rstrip', () => {
-    const _strip = new Strip();
-    expect(_strip.rstrip('xyzaaa', 'a')).toBe('xyz');
-    expect(_strip.rstrip('xyzaaa', 'b')).toBe('xyzaaa');
-    expect(_strip.rstrip('xyzaba', 'b')).toBe('xyzaba');
-    expect(_strip.rstrip('xyzbba', 'b')).toBe('xyzbba');
-    expect(_strip.rstrip('bbaxyzbbb', 'b')).toBe('bbaxyz');
-    expect(_strip.rstrip('bbaxyz   ')).toBe('bbaxyz');
-    expect(_strip.rstrip('bb   axyz   ')).toBe('bb   axyz');
-    expect(_strip.rstrip('   bb   axyz   ')).toBe('   bb   axyz');
-})
-
-
-test('Strip strip', () => {
-    const _strip = new Strip();
-    expect(_strip.strip('xyzaaa', 'a')).toBe('xyz');
-    expect(_strip.strip('xyzaaa', 'b')).toBe('xyzaaa');
-    expect(_strip.strip('xyzaba', 'b')).toBe('xyzaba');
-    expect(_strip.strip('xyzbba', 'b')).toBe('xyzbba');
-    expect(_strip.strip('bbaxyzbbb', 'b')).toBe('axyz');
-    expect(_strip.strip('bbabbaxyzbbabbb', 'b')).toBe('abbaxyzbba');
-    expect(_strip.strip('bbaxyz   ')).toBe('bbaxyz');
-    expect(_strip.strip('bb   axyz   ')).toBe('bb   axyz');
-    expect(_strip.strip('   bbaxyz')).toBe('bbaxyz');
-    expect(_strip.strip('   bb   axyz')).toBe('bb   axyz');
-    expect(_strip.strip('   bb   axyz   ')).toBe('bb   axyz');
 })
